Refactor Article service and DTO to improve data handling and structure

+++ src/main/java/cn/moonshotacademy/memoirs/entity/ArticleEntity.java
@@ -18,10 +18,10 @@ public class ArticleEntity {
     private String text;
 
     @Column(name = "teller_id")
-    private Long tellerId;
+    private Integer tellerId;
 
     @Column(name = "user_id")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "description")
     private String description;
